Move download button column into table-config
